@@ -1,12 +1,12 @@
 [project]
 name = "ConfigSpace"
-version = "0.7.2"
+version = "0.8.0"
 description = """\
     Creation and manipulation of parameter configuration spaces for \
     automated algorithm configuration and hyperparameter tuning. \
     """
 license.file = "LICENSE"
-requires-python = ">=3.7"
+requires-python = ">=3.8"
 readme = "README.md"
 authors = [
   { name = "Matthias Feurer" },
@@ -84,7 +84,7 @@ docs = [
 ]
 
 [build-system]
-requires = ["setuptools", "wheel", "oldest-supported-numpy", "Cython"]
+requires = ["setuptools", "wheel", "oldest-supported-numpy", "Cython==0.29.36"]
 build-backend = "setuptools.build_meta"
 
 
@@ -113,12 +113,12 @@ exclude_lines = [
 ]
 
 [tool.black]
-target-version = ['py37']
+target-version = ['py38']
 line-length = 100
 
 # https://github.com/charliermarsh/ruff
 [tool.ruff]
-target-version = "py37"
+target-version = "py38"
 line-length = 100
 show-source = true
 src = ["ConfigSpace", "test"]
@@ -274,7 +274,7 @@ convention = "numpy"
 
 
 [tool.mypy]
-python_version = "3.7"
+python_version = "3.8"
 packages = ["ConfigSpace", "test"]
 
 show_error_codes = true
