@@ -1,24 +1,29 @@
+from __future__ import annotations
+
 import io
 import warnings
-from typing import Any, Dict, Union, Optional
-
-from scipy.stats import beta as spbeta
+from typing import Any
 
 import numpy as np
-cimport numpy as np
-np.import_array()
+from scipy.stats import beta as spbeta
 
-from ConfigSpace.hyperparameters.beta_integer cimport BetaIntegerHyperparameter
+from ConfigSpace.hyperparameters.beta_integer import BetaIntegerHyperparameter
+from ConfigSpace.hyperparameters.uniform_float import UniformFloatHyperparameter
 
 
-cdef class BetaFloatHyperparameter(UniformFloatHyperparameter):
-
-    def __init__(self, name: str, alpha: Union[int, float], beta: Union[int, float],
-                 lower: Union[float, int],
-                 upper: Union[float, int],
-                 default_value: Union[None, float] = None,
-                 q: Union[int, float, None] = None, log: bool = False,
-                 meta: Optional[Dict] = None) -> None:
+class BetaFloatHyperparameter(UniformFloatHyperparameter):
+    def __init__(
+        self,
+        name: str,
+        alpha: int | float,
+        beta: int | float,
+        lower: float | int,
+        upper: float | int,
+        default_value: None | float = None,
+        q: int | float | None = None,
+        log: bool = False,
+        meta: dict | None = None,
+    ) -> None:
         r"""
         A beta distributed float hyperparameter. The 'lower' and 'upper' parameters move the
         distribution from the [0, 1]-range and scale it appropriately, but the shape of the
@@ -62,24 +67,38 @@ cdef class BetaFloatHyperparameter(UniformFloatHyperparameter):
         # error, which we do by setting default_value = upper - lower / 2 to not raise an error,
         # then actually call check_default once we have alpha and beta, and are not inside
         # UniformFloatHP.
-        super(BetaFloatHyperparameter, self).__init__(
-            name, lower, upper, (upper + lower) / 2, q, log, meta)
+        super().__init__(
+            name, lower, upper, (upper + lower) / 2, q, log, meta,
+        )
         self.alpha = float(alpha)
         self.beta = float(beta)
         if (alpha < 1) or (beta < 1):
-            raise ValueError("Please provide values of alpha and beta larger than or equal to\
-             1 so that the probability density is finite.")
+            raise ValueError(
+                "Please provide values of alpha and beta larger than or equal to\
+             1 so that the probability density is finite.",
+            )
 
         if (self.q is not None) and (self.log is not None) and (default_value is None):
-            warnings.warn("Logscale and quantization together results in incorrect default values. "
-                          "We recommend specifying a default value manually for this specific case.")
+            warnings.warn(
+                "Logscale and quantization together results in incorrect default values. "
+                "We recommend specifying a default value manually for this specific case.",
+            )
 
         self.default_value = self.check_default(default_value)
         self.normalized_default_value = self._inverse_transform(self.default_value)
 
     def __repr__(self) -> str:
         repr_str = io.StringIO()
-        repr_str.write("%s, Type: BetaFloat, Alpha: %s Beta: %s, Range: [%s, %s], Default: %s" % (self.name, repr(self.alpha), repr(self.beta), repr(self.lower), repr(self.upper), repr(self.default_value)))
+        repr_str.write(
+            "{}, Type: BetaFloat, Alpha: {} Beta: {}, Range: [{}, {}], Default: {}".format(
+                self.name,
+                repr(self.alpha),
+                repr(self.beta),
+                repr(self.lower),
+                repr(self.upper),
+                repr(self.default_value),
+            ),
+        )
 
         if self.log:
             repr_str.write(", on log-scale")
@@ -105,14 +124,14 @@ cdef class BetaFloatHyperparameter(UniformFloatHyperparameter):
             return False
 
         return (
-            self.name == other.name and
-            self.default_value == other.default_value and
-            self.alpha == other.alpha and
-            self.beta == other.beta and
-            self.log == other.log and
-            self.q == other.q and
-            self.lower == other.lower and
-            self.upper == other.upper
+            self.name == other.name
+            and self.default_value == other.default_value
+            and self.alpha == other.alpha
+            and self.beta == other.beta
+            and self.log == other.log
+            and self.q == other.q
+            and self.lower == other.lower
+            and self.upper == other.upper
         )
 
     def __copy__(self):
@@ -125,20 +144,24 @@ cdef class BetaFloatHyperparameter(UniformFloatHyperparameter):
             q=self.q,
             lower=self.lower,
             upper=self.upper,
-            meta=self.meta
+            meta=self.meta,
         )
 
     def __hash__(self):
         return hash((self.name, self.alpha, self.beta, self.lower, self.upper, self.log, self.q))
 
-    def to_uniform(self) -> "UniformFloatHyperparameter":
-        return UniformFloatHyperparameter(self.name,
-                                          self.lower,
-                                          self.upper,
-                                          default_value=self.default_value,
-                                          q=self.q, log=self.log, meta=self.meta)
+    def to_uniform(self) -> UniformFloatHyperparameter:
+        return UniformFloatHyperparameter(
+            self.name,
+            self.lower,
+            self.upper,
+            default_value=self.default_value,
+            q=self.q,
+            log=self.log,
+            meta=self.meta,
+        )
 
-    def check_default(self, default_value: Union[int, float, None]) -> Union[int, float]:
+    def check_default(self, default_value: int | float | None) -> int | float:
         # return mode as default
         # TODO - for log AND quantization together specifially, this does not give the exact right
         # value, due to the bounds _lower and _upper being adjusted when quantizing in
@@ -160,28 +183,36 @@ cdef class BetaFloatHyperparameter(UniformFloatHyperparameter):
         else:
             raise ValueError("Illegal default value %s" % str(default_value))
 
-    def to_integer(self) -> "BetaIntegerHyperparameter":
-        if self.q is None:
-            q_int = None
-        else:
-            q_int = int(np.rint(self.q))
+    def to_integer(self) -> BetaIntegerHyperparameter:
+        q_int = None if self.q is None else int(np.rint(self.q))
 
         lower = int(np.ceil(self.lower))
         upper = int(np.floor(self.upper))
         default_value = int(np.rint(self.default_value))
-        return BetaIntegerHyperparameter(self.name, lower=lower, upper=upper, alpha=self.alpha, beta=self.beta,
-                                         default_value=default_value, q=q_int, log=self.log)
+        return BetaIntegerHyperparameter(
+            self.name,
+            lower=lower,
+            upper=upper,
+            alpha=self.alpha,
+            beta=self.beta,
+            default_value=default_value,
+            q=q_int,
+            log=self.log,
+        )
 
-    def is_legal(self, value: Union[float]) -> bool:
+    def is_legal(self, value: float) -> bool:
         if isinstance(value, (float, int)):
             return self.upper >= value >= self.lower
         return False
 
-    cpdef bint is_legal_vector(self, DTYPE_t value):
+    def is_legal_vector(self, value) -> bool:
         return self._upper >= value >= self._lower
 
-    def _sample(self, rs: np.random.RandomState, size: Optional[int] = None
-                ) -> Union[np.ndarray, float]:
+    def _sample(
+        self,
+        rs: np.random.RandomState,
+        size: int | None = None,
+    ) -> np.ndarray | float:
         alpha = self.alpha
         beta = self.beta
         return spbeta(alpha, beta).rvs(size=size, random_state=rs)
@@ -201,7 +232,7 @@ cdef class BetaFloatHyperparameter(UniformFloatHyperparameter):
             function is to be computed.
 
         Returns
-        ----------
+        -------
         np.ndarray(N, )
             Probability density values of the input vector
         """
@@ -209,8 +240,11 @@ cdef class BetaFloatHyperparameter(UniformFloatHyperparameter):
         lb = self._inverse_transform(self.lower)
         alpha = self.alpha
         beta = self.beta
-        return spbeta(alpha, beta, loc=lb, scale=ub-lb).pdf(vector) \
-            * (ub-lb) / (self._upper - self._lower)
+        return (
+            spbeta(alpha, beta, loc=lb, scale=ub - lb).pdf(vector)
+            * (ub - lb)
+            / (self._upper - self._lower)
+        )
 
     def get_max_density(self) -> float:
         if (self.alpha > 1) or (self.beta > 1):
