@@ -1,11 +1,9 @@
 from typing import Dict, Optional, Union
 
 import numpy as np
-cimport numpy as np
-np.import_array()
 
 
-cdef class Hyperparameter(object):
+class Hyperparameter:
 
     def __init__(self, name: str, meta: Optional[Dict]) -> None:
         if not isinstance(name, str):
@@ -21,7 +19,7 @@ cdef class Hyperparameter(object):
     def is_legal(self, value):
         raise NotImplementedError()
 
-    cpdef bint is_legal_vector(self, DTYPE_t value):
+    def is_legal_vector(self, value) -> int:
         """
         Check whether the given value is a legal value for the vector
         representation of this hyperparameter.
@@ -46,7 +44,7 @@ cdef class Hyperparameter(object):
     def rvs(
         self,
         size: Optional[int] = None,
-        random_state: Optional[Union[int, np.random, np.random.RandomState]] = None
+        random_state: Optional[Union[int, np.random, np.random.RandomState]] = None,
     ) -> Union[float, np.ndarray]:
         """
         scipy compatibility wrapper for ``_sample``,
@@ -86,7 +84,7 @@ cdef class Hyperparameter(object):
 
         vector = self._sample(
             rs=check_random_state(random_state),
-            size=size if size is not None else 1
+            size=size if size is not None else 1,
         )
         if size is None:
             vector = vector[0]
@@ -98,7 +96,7 @@ cdef class Hyperparameter(object):
 
     def _transform(
         self,
-        vector: Union[np.ndarray, float, int]
+        vector: Union[np.ndarray, float, int],
     ) -> Optional[Union[np.ndarray, float, int]]:
         raise NotImplementedError()
 
@@ -114,7 +112,7 @@ cdef class Hyperparameter(object):
     def get_num_neighbors(self, value):
         raise NotImplementedError()
 
-    cpdef int compare_vector(self, DTYPE_t value, DTYPE_t value2):
+    def compare_vector(self, DTYPE_t value, DTYPE_t value2) -> int:
         raise NotImplementedError()
 
     def pdf(self, vector: np.ndarray) -> np.ndarray:
