@@ -1,19 +1,26 @@
+from __future__ import annotations
+
 import io
-from typing import Dict, List, Optional, Union
 import warnings
 
 import numpy as np
-cimport numpy as np
-np.import_array()
 
 from ConfigSpace.functional import center_range
-from ConfigSpace.hyperparameters.uniform_float cimport UniformFloatHyperparameter
+from ConfigSpace.hyperparameters.integer_hyperparameter import IntegerHyperparameter
+from ConfigSpace.hyperparameters.uniform_float import UniformFloatHyperparameter
 
 
-cdef class UniformIntegerHyperparameter(IntegerHyperparameter):
-    def __init__(self, name: str, lower: int, upper: int, default_value: Union[int, None] = None,
-                 q: Union[int, None] = None, log: bool = False,
-                 meta: Optional[Dict] = None) -> None:
+class UniformIntegerHyperparameter(IntegerHyperparameter):
+    def __init__(
+        self,
+        name: str,
+        lower: int,
+        upper: int,
+        default_value: int | None = None,
+        q: int | None = None,
+        log: bool = False,
+        meta: dict | None = None,
+    ) -> None:
         """
         A uniformly distributed integer hyperparameter.
 
@@ -44,54 +51,61 @@ cdef class UniformIntegerHyperparameter(IntegerHyperparameter):
             Field for holding meta data provided by the user.
             Not used by the configuration space.
         """
-
-        super(UniformIntegerHyperparameter, self).__init__(name, default_value, meta)
         self.lower = self.check_int(lower, "lower")
         self.upper = self.check_int(upper, "upper")
+
         if default_value is not None:
             default_value = self.check_int(default_value, name)
+        else:
+            default_value = self.check_default(default_value)
+
+        # NOTE: Placed after the default value check to ensure it's set and not None
+        super().__init__(name, default_value, meta)
 
         if q is not None:
             if q < 1:
-                warnings.warn("Setting quantization < 1 for Integer "
-                              "Hyperparameter '%s' has no effect." %
-                              name)
+                warnings.warn(
+                    "Setting quantization < 1 for Integer "
+                    "Hyperparameter '%s' has no effect." % name,
+                )
                 self.q = None
             else:
                 self.q = self.check_int(q, "q")
                 if (self.upper - self.lower) % self.q != 0:
                     raise ValueError(
                         "Upper bound (%d) - lower bound (%d) must be a multiple of q (%d)"
-                        % (self.upper, self.lower, self.q)
+                        % (self.upper, self.lower, self.q),
                     )
         else:
             self.q = None
         self.log = bool(log)
 
         if self.lower >= self.upper:
-            raise ValueError("Upper bound %d must be larger than lower bound "
-                             "%d for hyperparameter %s" %
-                             (self.lower, self.upper, name))
+            raise ValueError(
+                "Upper bound %d must be larger than lower bound "
+                "%d for hyperparameter %s" % (self.lower, self.upper, name),
+            )
         elif log and self.lower <= 0:
-            raise ValueError("Negative lower bound (%d) for log-scale "
-                             "hyperparameter %s is forbidden." %
-                             (self.lower, name))
+            raise ValueError(
+                "Negative lower bound (%d) for log-scale "
+                "hyperparameter %s is forbidden." % (self.lower, name),
+            )
 
-        self.default_value = self.check_default(default_value)
-
-        self.ufhp = UniformFloatHyperparameter(self.name,
-                                               self.lower - 0.49999,
-                                               self.upper + 0.49999,
-                                               log=self.log,
-                                               default_value=self.default_value)
+        self.ufhp = UniformFloatHyperparameter(
+            self.name,
+            self.lower - 0.49999,
+            self.upper + 0.49999,
+            log=self.log,
+            default_value=self.default_value,
+        )
 
         self.normalized_default_value = self._inverse_transform(self.default_value)
 
     def __repr__(self) -> str:
         repr_str = io.StringIO()
-        repr_str.write("%s, Type: UniformInteger, Range: [%s, %s], Default: %s"
-                       % (self.name, repr(self.lower),
-                          repr(self.upper), repr(self.default_value)))
+        repr_str.write(
+            f"{self.name}, Type: UniformInteger, Range: [{self.lower!r}, {self.upper!r}], Default: {self.default_value!r}",
+        )
         if self.log:
             repr_str.write(", on log-scale")
         if self.q is not None:
@@ -99,17 +113,19 @@ cdef class UniformIntegerHyperparameter(IntegerHyperparameter):
         repr_str.seek(0)
         return repr_str.getvalue()
 
-    def _sample(self, rs: np.random.RandomState, size: Optional[int] = None
-                ) -> Union[np.ndarray, float]:
+    def _sample(
+        self,
+        rs: np.random.RandomState,
+        size: int | None = None,
+    ) -> np.ndarray | float:
         value = self.ufhp._sample(rs, size=size)
         # Map all floats which belong to the same integer value to the same
         # float value by first transforming it to an integer and then
         # transforming it back to a float between zero and one
         value = self._transform(value)
-        value = self._inverse_transform(value)
-        return value
+        return self._inverse_transform(value)
 
-    cpdef np.ndarray _transform_vector(self, np.ndarray vector):
+    def _transform_vector(self, vector: np.ndarray) -> np.ndarray:
         vector = self.ufhp._transform_vector(vector)
         if self.q is not None:
             vector = np.rint((vector - self.lower) / self.q) * self.q + self.lower
@@ -118,7 +134,7 @@ cdef class UniformIntegerHyperparameter(IntegerHyperparameter):
 
         return np.rint(vector)
 
-    cpdef long long _transform_scalar(self, double scalar):
+    def _transform_scalar(self, scalar: float) -> float:
         scalar = self.ufhp._transform_scalar(scalar)
         if self.q is not None:
             scalar = np.round((scalar - self.lower) / self.q) * self.q + self.lower
@@ -126,30 +142,26 @@ cdef class UniformIntegerHyperparameter(IntegerHyperparameter):
             scalar = max(scalar, self.lower)
         return int(np.round(scalar))
 
-    def _inverse_transform(self, vector: Union[np.ndarray, float, int]
-                           ) -> Union[np.ndarray, float, int]:
+    def _inverse_transform(
+        self,
+        vector: np.ndarray | float | int,
+    ) -> np.ndarray | float | int:
         return self.ufhp._inverse_transform(vector)
 
     def is_legal(self, value: int) -> bool:
         if not (isinstance(value, (int, np.int32, np.int64))):
             return False
-        elif self.upper >= value >= self.lower:
-            return True
-        else:
-            return False
+        return self.upper >= value >= self.lower
 
-    cpdef bint is_legal_vector(self, DTYPE_t value):
-        if 1.0 >= value >= 0.0:
-            return True
-        else:
-            return False
+    def is_legal_vector(self, value) -> int:
+        return 1.0 >= value >= 0.0
 
-    def check_default(self, default_value: Union[int, float]) -> int:
+    def check_default(self, default_value: int | float | None) -> int:
         if default_value is None:
             if self.log:
-                default_value = np.exp((np.log(self.lower) + np.log(self.upper)) / 2.)
+                default_value = np.exp((np.log(self.lower) + np.log(self.upper)) / 2.0)
             else:
-                default_value = (self.lower + self.upper) / 2.
+                default_value = (self.lower + self.upper) / 2.0
         default_value = int(np.round(default_value, 0))
 
         if self.is_legal(default_value):
@@ -166,12 +178,9 @@ cdef class UniformIntegerHyperparameter(IntegerHyperparameter):
             lower = self.ufhp._lower
 
         # If there is only one active value, this is not enough
-        if upper - lower >= 1:
-            return True
-        else:
-            return False
+        return upper - lower >= 1
 
-    def get_num_neighbors(self, value = None) -> int:
+    def get_num_neighbors(self, value=None) -> int:
         # If there is a value in the range, then that value is not a neighbor of itself
         # so we need to remove one
         if value is not None and self.lower <= value <= self.upper:
@@ -186,8 +195,8 @@ cdef class UniformIntegerHyperparameter(IntegerHyperparameter):
         number: int = 4,
         transform: bool = False,
         std: float = 0.2,
-    ) -> List[int]:
-        """Get the neighbors of a value
+    ) -> list[int]:
+        """Get the neighbors of a value.
 
         NOTE
         ----
@@ -223,22 +232,19 @@ cdef class UniformIntegerHyperparameter(IntegerHyperparameter):
             " if assumed to be in the unit-hypercube [0, 1]. If this was not"
             " the behaviour assumed, please raise a ticket on github."
         )
-        assert number < 1000000, (
-            "Can only generate less than 1 million neighbors."
-        )
+        assert number < 1000000, "Can only generate less than 1 million neighbors."
         # Convert python values to cython ones
-        cdef long long center = self._transform(value)
-        cdef long long lower = self.lower
-        cdef long long upper = self.upper
-        cdef unsigned int n_requested = number
-        cdef unsigned long long n_neighbors = upper - lower - 1
-        cdef long long stepsize = self.q if self.q is not None else 1
+        center = self._transform(value)
+        lower = self.lower
+        upper = self.upper
+        n_requested = number
+        n_neighbors = upper - lower - 1
+        stepsize = self.q if self.q is not None else 1
 
         neighbors = []
 
-        cdef long long v  # A value that's possible to return
+        v: int  # A value that's possible to return
         if n_neighbors < n_requested:
-
             for v in range(lower, center):
                 neighbors.append(v)
 
@@ -254,29 +260,24 @@ cdef class UniformIntegerHyperparameter(IntegerHyperparameter):
         # This will be sampled from and converted to the corresponding int value
         # However, this is too slow - we use the "poor man's truncnorm below"
         # cdef np.ndarray float_indices = truncnorm.rvs(
-        #     a=(0 - value) / std,
-        #     b=(1 - value) / std,
-        #     loc=value,
-        #     scale=std,
-        #     size=number,
-        #     random_state=rs
-        # )
         # We sample five times as many values as needed and weed them out below
         # (perform rejection sampling and make sure we don't sample any neighbor twice)
         # This increases our chances of not having to fill the neighbors list by calling
         # `center_range`
         # Five is an arbitrary number and can probably be tuned to reduce overhead
-        cdef np.ndarray float_indices = rs.normal(value, std, size=number * 5)
-        cdef np.ndarray mask = (float_indices >= 0) & (float_indices <= 1)
+        float_indices: np.ndarray = rs.normal(value, std, size=number * 5)
+        mask: np.ndarray = (float_indices >= 0) & (float_indices <= 1)
         float_indices = float_indices[mask]
 
-        cdef np.ndarray possible_neighbors_as_array = self._transform_vector(float_indices).astype(np.longlong)
-        cdef long long [:] possible_neighbors = possible_neighbors_as_array
+        possible_neighbors_as_array: np.ndarray = self._transform_vector(float_indices).astype(
+            np.longlong,
+        )
+        possible_neighbors: np.ndarray = possible_neighbors_as_array
 
-        cdef unsigned int n_neighbors_generated = 0
-        cdef unsigned int n_candidates = len(float_indices)
-        cdef unsigned int candidate_index = 0
-        cdef set seen = {center}
+        n_neighbors_generated: int = 0
+        n_candidates: int = len(float_indices)
+        candidate_index: int = 0
+        seen: set[int] = {center}
         while n_neighbors_generated < n_requested and candidate_index < n_candidates:
             v = possible_neighbors[candidate_index]
             if v not in seen:
@@ -308,7 +309,7 @@ cdef class UniformIntegerHyperparameter(IntegerHyperparameter):
         distributions, only normal distributions (as the inverse_transform
         in the pdf method handles these). Optimally, an IntegerHyperparameter
         should have a corresponding float, which can be utlized for the calls
-        to the probability density function (see e.g. NormalIntegerHyperparameter)
+        to the probability density function (see e.g. NormalIntegerHyperparameter).
 
         Parameters
         ----------
@@ -317,7 +318,7 @@ cdef class UniformIntegerHyperparameter(IntegerHyperparameter):
             function is to be computed.
 
         Returns
-        ----------
+        -------
         np.ndarray(N, )
             Probability density values of the input vector
         """
@@ -329,8 +330,5 @@ cdef class UniformIntegerHyperparameter(IntegerHyperparameter):
         return 1 / (ub - lb + 1)
 
     def get_size(self) -> float:
-        if self.q is None:
-            q = 1
-        else:
-            q = self.q
+        q = 1 if self.q is None else self.q
         return np.rint((self.upper - self.lower) / q) + 1
