@@ -1140,8 +1140,8 @@ class ConfigurationSpace(Mapping[str, Hyperparameter]):
 
         return hp
 
-    def __contains__(self, key: str) -> bool:
-        return key in self._hyperparameters
+    #def __contains__(self, key: str) -> bool:
+    #    return key in self._hyperparameters
 
     def __repr__(self) -> str:
         retval = io.StringIO()
