@@ -1,27 +1,19 @@
-from typing import Any, Dict, List, Optional, Union
+from __future__ import annotations
+
+from typing import Any, Optional, Union
 
 import numpy as np
-cimport numpy as np
-np.import_array()
 
-# We now need to fix a datatype for our arrays. I've used the variable
-# DTYPE for this, which is assigned to the usual NumPy runtime
-# type info object.
-DTYPE = float
-# "ctypedef" assigns a corresponding compile-time type to DTYPE_t. For
-# every type in the numpy module there's a corresponding compile-time
-# type with a _t-suffix.
-ctypedef np.float_t DTYPE_t
-
-from ConfigSpace.hyperparameters.hyperparameter cimport Hyperparameter
+from ConfigSpace.hyperparameters.hyperparameter import Hyperparameter
 
 
-cdef class Constant(Hyperparameter):
-    cdef public value
-    cdef DTYPE_t value_vector
-
-    def __init__(self, name: str, value: Union[str, int, float], meta: Optional[Dict] = None
-                 ) -> None:
+class Constant(Hyperparameter):
+    def __init__(
+        self,
+        name: str,
+        value: Union[str, int, float],
+        meta: Optional[dict] = None,
+    ) -> None:
         """
         Representing a constant hyperparameter in the configuration space.
 
@@ -41,21 +33,19 @@ cdef class Constant(Hyperparameter):
         super(Constant, self).__init__(name, meta)
         allowed_types = (int, float, str)
 
-        if not isinstance(value, allowed_types) or \
-                isinstance(value, bool):
-            raise TypeError("Constant value is of type %s, but only the "
-                            "following types are allowed: %s" %
-                            (type(value), allowed_types))  # type: ignore
+        if not isinstance(value, allowed_types) or isinstance(value, bool):
+            raise TypeError(
+                "Constant value is of type %s, but only the "
+                "following types are allowed: %s" % (type(value), allowed_types),
+            )  # type: ignore
 
         self.value = value
-        self.value_vector = 0.
+        self.value_vector = 0.0
         self.default_value = value
-        self.normalized_default_value = 0.
+        self.normalized_default_value = 0.0
 
     def __repr__(self) -> str:
-        repr_str = ["%s" % self.name,
-                    "Type: Constant",
-                    "Value: %s" % self.value]
+        repr_str = ["%s" % self.name, "Type: Constant", "Value: %s" % self.value]
         return ", ".join(repr_str)
 
     def __eq__(self, other: Any) -> bool:
@@ -75,9 +65,9 @@ cdef class Constant(Hyperparameter):
             return False
 
         return (
-            self.value == other.value and
-            self.name == other.name and
-            self.default_value == other.default_value
+            self.value == other.value
+            and self.name == other.name
+            and self.default_value == other.default_value
         )
 
     def __copy__(self):
@@ -89,26 +79,31 @@ cdef class Constant(Hyperparameter):
     def is_legal(self, value: Union[str, int, float]) -> bool:
         return value == self.value
 
-    cpdef bint is_legal_vector(self, DTYPE_t value):
+    def is_legal_vector(self, value) -> int:
         return value == self.value_vector
 
     def _sample(self, rs: None, size: Optional[int] = None) -> Union[int, np.ndarray]:
         return 0 if size == 1 else np.zeros((size,))
 
-    def _transform(self, vector: Optional[Union[np.ndarray, float, int]]) \
-            -> Optional[Union[np.ndarray, float, int]]:
+    def _transform(
+        self, vector: Optional[Union[np.ndarray, float, int]],
+    ) -> Optional[Union[np.ndarray, float, int]]:
         return self.value
 
-    def _transform_vector(self, vector: Optional[np.ndarray]) \
-            -> Optional[Union[np.ndarray, float, int]]:
+    def _transform_vector(
+        self, vector: Optional[np.ndarray],
+    ) -> Optional[Union[np.ndarray, float, int]]:
         return self.value
 
-    def _transform_scalar(self, vector: Optional[Union[float, int]]) \
-            -> Optional[Union[np.ndarray, float, int]]:
+    def _transform_scalar(
+        self, vector: Optional[Union[float, int]],
+    ) -> Optional[Union[np.ndarray, float, int]]:
         return self.value
 
-    def _inverse_transform(self, vector: Union[np.ndarray, float, int]
-                           ) -> Union[np.ndarray, int, float]:
+    def _inverse_transform(
+        self,
+        vector: Union[np.ndarray, float, int],
+    ) -> Union[np.ndarray, int, float]:
         if vector != self.value:
             return np.NaN
         return 0
@@ -116,11 +111,12 @@ cdef class Constant(Hyperparameter):
     def has_neighbors(self) -> bool:
         return False
 
-    def get_num_neighbors(self, value = None) -> int:
+    def get_num_neighbors(self, value=None) -> int:
         return 0
 
-    def get_neighbors(self, value: Any, rs: np.random.RandomState, number: int,
-                      transform: bool = False) -> List:
+    def get_neighbors(
+        self, value: Any, rs: np.random.RandomState, number: int, transform: bool = False,
+    ) -> list:
         return []
 
     def pdf(self, vector: np.ndarray) -> np.ndarray:
@@ -139,7 +135,7 @@ cdef class Constant(Hyperparameter):
             function is to be computed.
 
         Returns
-        ----------
+        -------
         np.ndarray(N, )
             Probability density values of the input vector
         """
@@ -162,7 +158,7 @@ cdef class Constant(Hyperparameter):
             function is to be computed.
 
         Returns
-        ----------
+        -------
         np.ndarray(N, )
             Probability density values of the input vector
         """
@@ -174,5 +170,6 @@ cdef class Constant(Hyperparameter):
     def get_size(self) -> float:
         return 1.0
 
-cdef class UnParametrizedHyperparameter(Constant):
+
+class UnParametrizedHyperparameter(Constant):
     pass
