@@ -27,9 +27,8 @@
 # SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 from __future__ import annotations
 
-import unittest
-
 import numpy as np
+import pytest
 
 from ConfigSpace.conditions import (
     AndConjunction,
@@ -49,332 +48,306 @@ from ConfigSpace.hyperparameters import (
 )
 
 
-class TestConditions(unittest.TestCase):
-    # TODO: return only copies of the objects!
-    def test_equals_condition(self):
-        hp1 = CategoricalHyperparameter("parent", [0, 1])
-        hp2 = UniformIntegerHyperparameter("child", 0, 10)
-        cond = EqualsCondition(hp2, hp1, 0)
-        cond_ = EqualsCondition(hp2, hp1, 0)
+# TODO: return only copies of the objects!
+def test_equals_condition():
+    hp1 = CategoricalHyperparameter("parent", [0, 1])
+    hp2 = UniformIntegerHyperparameter("child", 0, 10)
+    cond = EqualsCondition(hp2, hp1, 0)
+    cond_ = EqualsCondition(hp2, hp1, 0)
 
-        # Test vector value:
-        assert cond.vector_value == hp1._inverse_transform(0)
-        assert cond.vector_value == cond_.vector_value
+    # Test vector value:
+    assert cond.vector_value == hp1._inverse_transform(0)
+    assert cond.vector_value == cond_.vector_value
 
-        # Test invalid conditions:
-        self.assertRaises(TypeError, EqualsCondition, hp2, "parent", 0)
-        self.assertRaises(TypeError, EqualsCondition, "child", hp1, 0)
-        self.assertRaises(ValueError, EqualsCondition, hp1, hp1, 0)
+    # Test invalid conditions:
+    with pytest.raises(TypeError):
+        EqualsCondition(hp2, "parent", 0)
+    with pytest.raises(TypeError):
+        EqualsCondition("child", hp1, 0)
+    with pytest.raises(ValueError):
+        EqualsCondition(hp1, hp1, 0)
 
-        assert cond == cond_
+    assert cond == cond_
 
-        cond_reverse = EqualsCondition(hp1, hp2, 0)
-        assert cond != cond_reverse
+    cond_reverse = EqualsCondition(hp1, hp2, 0)
+    assert cond != cond_reverse
 
-        assert cond != {}
+    assert cond != {}
 
-        assert str(cond) == "child | parent == 0"
+    assert str(cond) == "child | parent == 0"
 
-    def test_equals_condition_illegal_value(self):
-        epsilon = UniformFloatHyperparameter("epsilon", 1e-5, 1e-1, default_value=1e-4, log=True)
-        loss = CategoricalHyperparameter(
-            "loss",
-            ["hinge", "log", "modified_huber", "squared_hinge", "perceptron"],
-            default_value="hinge",
-        )
-        self.assertRaisesRegex(
-            ValueError,
-            "Hyperparameter 'epsilon' is "
-            "conditional on the illegal value 'huber' of "
-            "its parent hyperparameter 'loss'",
-            EqualsCondition,
-            epsilon,
-            loss,
-            "huber",
-        )
 
-    def test_not_equals_condition(self):
-        hp1 = CategoricalHyperparameter("parent", [0, 1])
-        hp2 = UniformIntegerHyperparameter("child", 0, 10)
-        cond = NotEqualsCondition(hp2, hp1, 0)
-        cond_ = NotEqualsCondition(hp2, hp1, 0)
-        assert cond == cond_
+def test_equals_condition_illegal_value():
+    epsilon = UniformFloatHyperparameter("epsilon", 1e-5, 1e-1, default_value=1e-4, log=True)
+    loss = CategoricalHyperparameter(
+        "loss",
+        ["hinge", "log", "modified_huber", "squared_hinge", "perceptron"],
+        default_value="hinge",
+    )
+    with pytest.raises(ValueError):
+        EqualsCondition(epsilon, loss, "huber")
 
-        # Test vector value:
-        assert cond.vector_value == hp1._inverse_transform(0)
-        assert cond.vector_value == cond_.vector_value
 
-        cond_reverse = NotEqualsCondition(hp1, hp2, 0)
-        assert cond != cond_reverse
+def test_not_equals_condition():
+    hp1 = CategoricalHyperparameter("parent", [0, 1])
+    hp2 = UniformIntegerHyperparameter("child", 0, 10)
+    cond = NotEqualsCondition(hp2, hp1, 0)
+    cond_ = NotEqualsCondition(hp2, hp1, 0)
+    assert cond == cond_
 
-        assert cond != {}
+    # Test vector value:
+    assert cond.vector_value == hp1._inverse_transform(0)
+    assert cond.vector_value == cond_.vector_value
 
-        assert str(cond) == "child | parent != 0"
+    cond_reverse = NotEqualsCondition(hp1, hp2, 0)
+    assert cond != cond_reverse
 
-    def test_not_equals_condition_illegal_value(self):
-        epsilon = UniformFloatHyperparameter("epsilon", 1e-5, 1e-1, default_value=1e-4, log=True)
-        loss = CategoricalHyperparameter(
-            "loss",
-            ["hinge", "log", "modified_huber", "squared_hinge", "perceptron"],
-            default_value="hinge",
-        )
-        self.assertRaisesRegex(
-            ValueError,
-            "Hyperparameter 'epsilon' is "
-            "conditional on the illegal value 'huber' of "
-            "its parent hyperparameter 'loss'",
-            NotEqualsCondition,
-            epsilon,
-            loss,
-            "huber",
-        )
+    assert cond != {}
 
-    def test_in_condition(self):
-        hp1 = CategoricalHyperparameter("parent", list(range(0, 11)))
-        hp2 = UniformIntegerHyperparameter("child", 0, 10)
-        cond = InCondition(hp2, hp1, [0, 1, 2, 3, 4, 5])
-        cond_ = InCondition(hp2, hp1, [0, 1, 2, 3, 4, 5])
-        assert cond == cond_
+    assert str(cond) == "child | parent != 0"
 
-        # Test vector value:
-        assert cond.vector_values == [hp1._inverse_transform(i) for i in [0, 1, 2, 3, 4, 5]]
-        assert cond.vector_values == cond_.vector_values
 
-        cond_reverse = InCondition(hp1, hp2, [0, 1, 2, 3, 4, 5])
-        assert cond != cond_reverse
+def test_not_equals_condition_illegal_value():
+    epsilon = UniformFloatHyperparameter("epsilon", 1e-5, 1e-1, default_value=1e-4, log=True)
+    loss = CategoricalHyperparameter(
+        "loss",
+        ["hinge", "log", "modified_huber", "squared_hinge", "perceptron"],
+        default_value="hinge",
+    )
+    with pytest.raises(ValueError):
+        NotEqualsCondition(epsilon, loss, "huber")
 
-        assert cond != {}
 
-        assert str(cond) == "child | parent in {0, 1, 2, 3, 4, 5}"
+def test_in_condition():
+    hp1 = CategoricalHyperparameter("parent", list(range(11)))
+    hp2 = UniformIntegerHyperparameter("child", 0, 10)
+    cond = InCondition(hp2, hp1, [0, 1, 2, 3, 4, 5])
+    cond_ = InCondition(hp2, hp1, [0, 1, 2, 3, 4, 5])
+    assert cond == cond_
 
-    def test_greater_and_less_condition(self):
-        child = Constant("child", "child")
-        hp1 = UniformFloatHyperparameter("float", 0, 5)
-        hp2 = UniformIntegerHyperparameter("int", 0, 5)
-        hp3 = OrdinalHyperparameter("ord", list(range(6)))
+    # Test vector value:
+    assert cond.vector_values == [hp1._inverse_transform(i) for i in [0, 1, 2, 3, 4, 5]]
+    assert cond.vector_values == cond_.vector_values
 
-        for hp in [hp1, hp2, hp3]:
-            hyperparameter_idx = {child.name: 0, hp.name: 1}
+    cond_reverse = InCondition(hp1, hp2, [0, 1, 2, 3, 4, 5])
+    assert cond != cond_reverse
 
-            gt = GreaterThanCondition(child, hp, 1)
-            gt.set_vector_idx(hyperparameter_idx)
-            assert not gt.evaluate({hp.name: 0})
-            assert gt.evaluate({hp.name: 2})
-            assert not gt.evaluate({hp.name: None})
+    assert cond != {}
 
-            # Evaluate vector
-            test_value = hp._inverse_transform(2)
-            assert not gt.evaluate_vector(np.array([np.NaN, 0]))
-            assert gt.evaluate_vector(np.array([np.NaN, test_value]))
-            assert not gt.evaluate_vector(np.array([np.NaN, np.NaN]))
+    assert str(cond) == "child | parent in {0, 1, 2, 3, 4, 5}"
 
-            lt = LessThanCondition(child, hp, 1)
-            lt.set_vector_idx(hyperparameter_idx)
-            assert lt.evaluate({hp.name: 0})
-            assert not lt.evaluate({hp.name: 2})
-            assert not lt.evaluate({hp.name: None})
 
-            # Evaluate vector
-            test_value = hp._inverse_transform(2)
-            assert lt.evaluate_vector(np.array([np.NaN, 0, 0, 0]))
-            assert not lt.evaluate_vector(np.array([np.NaN, test_value]))
-            assert not lt.evaluate_vector(np.array([np.NaN, np.NaN]))
+def test_greater_and_less_condition():
+    child = Constant("child", "child")
+    hp1 = UniformFloatHyperparameter("float", 0, 5)
+    hp2 = UniformIntegerHyperparameter("int", 0, 5)
+    hp3 = OrdinalHyperparameter("ord", list(range(6)))
 
-        hp4 = CategoricalHyperparameter("cat", list(range(6)))
-        self.assertRaisesRegex(
-            ValueError,
-            "Parent hyperparameter in a > or < condition must be a subclass of "
-            "NumericalHyperparameter or OrdinalHyperparameter, but is "
-            "<cdef class 'ConfigSpace.hyperparameters.CategoricalHyperparameter'>",
-            GreaterThanCondition,
-            child,
-            hp4,
-            1,
-        )
-        self.assertRaisesRegex(
-            ValueError,
-            "Parent hyperparameter in a > or < condition must be a subclass of "
-            "NumericalHyperparameter or OrdinalHyperparameter, but is "
-            "<cdef class 'ConfigSpace.hyperparameters.CategoricalHyperparameter'>",
-            LessThanCondition,
-            child,
-            hp4,
-            1,
-        )
+    for hp in [hp1, hp2, hp3]:
+        hyperparameter_idx = {child.name: 0, hp.name: 1}
 
-        hp5 = OrdinalHyperparameter("ord", ["cold", "luke warm", "warm", "hot"])
-
-        hyperparameter_idx = {child.name: 0, hp5.name: 1}
-        gt = GreaterThanCondition(child, hp5, "warm")
+        gt = GreaterThanCondition(child, hp, 1)
         gt.set_vector_idx(hyperparameter_idx)
-        assert gt.evaluate({hp5.name: "hot"})
-        assert not gt.evaluate({hp5.name: "cold"})
+        assert not gt.evaluate({hp.name: 0})
+        assert gt.evaluate({hp.name: 2})
+        assert not gt.evaluate({hp.name: None})
 
-        assert gt.evaluate_vector(np.array([np.NaN, 3]))
+        # Evaluate vector
+        test_value = hp._inverse_transform(2)
         assert not gt.evaluate_vector(np.array([np.NaN, 0]))
+        assert gt.evaluate_vector(np.array([np.NaN, test_value]))
+        assert not gt.evaluate_vector(np.array([np.NaN, np.NaN]))
 
-        lt = LessThanCondition(child, hp5, "warm")
+        lt = LessThanCondition(child, hp, 1)
         lt.set_vector_idx(hyperparameter_idx)
-        assert lt.evaluate({hp5.name: "luke warm"})
-        assert not lt.evaluate({hp5.name: "warm"})
+        assert lt.evaluate({hp.name: 0})
+        assert not lt.evaluate({hp.name: 2})
+        assert not lt.evaluate({hp.name: None})
 
-        assert lt.evaluate_vector(np.array([np.NaN, 1]))
-        assert not lt.evaluate_vector(np.array([np.NaN, 2]))
+        # Evaluate vector
+        test_value = hp._inverse_transform(2)
+        assert lt.evaluate_vector(np.array([np.NaN, 0, 0, 0]))
+        assert not lt.evaluate_vector(np.array([np.NaN, test_value]))
+        assert not lt.evaluate_vector(np.array([np.NaN, np.NaN]))
 
-    def test_in_condition_illegal_value(self):
-        epsilon = UniformFloatHyperparameter("epsilon", 1e-5, 1e-1, default_value=1e-4, log=True)
-        loss = CategoricalHyperparameter(
-            "loss",
-            ["hinge", "log", "modified_huber", "squared_hinge", "perceptron"],
-            default_value="hinge",
-        )
-        self.assertRaisesRegex(
-            ValueError,
-            "Hyperparameter 'epsilon' is "
-            "conditional on the illegal value 'huber' of "
-            "its parent hyperparameter 'loss'",
-            InCondition,
-            epsilon,
-            loss,
-            ["huber", "log"],
-        )
+    hp4 = CategoricalHyperparameter("cat", list(range(6)))
+    with pytest.raises(ValueError):
+        GreaterThanCondition(child, hp4, 1)
 
-    def test_and_conjunction(self):
-        self.assertRaises(TypeError, AndConjunction, "String1", "String2")
+    with pytest.raises(ValueError):
+        LessThanCondition(child, hp4, 1)
 
-        hp1 = CategoricalHyperparameter("input1", [0, 1])
-        hp2 = CategoricalHyperparameter("input2", [0, 1])
-        hp3 = CategoricalHyperparameter("input3", [0, 1])
-        hp4 = Constant("And", "True")
-        cond1 = EqualsCondition(hp4, hp1, 1)
+    hp5 = OrdinalHyperparameter("ord", ["cold", "luke warm", "warm", "hot"])
 
-        # Only one condition in an AndConjunction!
-        self.assertRaises(ValueError, AndConjunction, cond1)
+    hyperparameter_idx = {child.name: 0, hp5.name: 1}
+    gt = GreaterThanCondition(child, hp5, "warm")
+    gt.set_vector_idx(hyperparameter_idx)
+    assert gt.evaluate({hp5.name: "hot"})
+    assert not gt.evaluate({hp5.name: "cold"})
 
-        cond2 = EqualsCondition(hp4, hp2, 1)
-        cond3 = EqualsCondition(hp4, hp3, 1)
+    assert gt.evaluate_vector(np.array([np.NaN, 3]))
+    assert not gt.evaluate_vector(np.array([np.NaN, 0]))
 
-        andconj1 = AndConjunction(cond1, cond2)
-        andconj1_ = AndConjunction(cond1, cond2)
-        assert andconj1 == andconj1_
+    lt = LessThanCondition(child, hp5, "warm")
+    lt.set_vector_idx(hyperparameter_idx)
+    assert lt.evaluate({hp5.name: "luke warm"})
+    assert not lt.evaluate({hp5.name: "warm"})
 
-        # Test setting vector idx
-        hyperparameter_idx = {hp1.name: 0, hp2.name: 1, hp3.name: 2, hp4.name: 3}
-        andconj1.set_vector_idx(hyperparameter_idx)
-        assert andconj1.get_parents_vector() == [0, 1]
-        assert andconj1.get_children_vector() == [3, 3]
+    assert lt.evaluate_vector(np.array([np.NaN, 1]))
+    assert not lt.evaluate_vector(np.array([np.NaN, 2]))
 
-        andconj2 = AndConjunction(cond2, cond3)
-        assert andconj1 != andconj2
 
-        andconj3 = AndConjunction(cond1, cond2, cond3)
-        assert str(andconj3) == "(And | input1 == 1 && And | input2 == 1 && And | input3 == 1)"
+def test_in_condition_illegal_value():
+    epsilon = UniformFloatHyperparameter("epsilon", 1e-5, 1e-1, default_value=1e-4, log=True)
+    loss = CategoricalHyperparameter(
+        "loss",
+        ["hinge", "log", "modified_huber", "squared_hinge", "perceptron"],
+        default_value="hinge",
+    )
+    with pytest.raises(ValueError):
+        InCondition(epsilon, loss, ["huber", "log"])
 
-        # Test __eq__
-        assert andconj1 != andconj3
-        assert andconj1 != "String"
 
-    def test_or_conjunction(self):
-        self.assertRaises(TypeError, AndConjunction, "String1", "String2")
+def test_and_conjunction():
+    with pytest.raises(TypeError):
+        AndConjunction("String1", "String2")
 
-        hp1 = CategoricalHyperparameter("input1", [0, 1])
-        hp2 = CategoricalHyperparameter("input2", [0, 1])
-        hp3 = CategoricalHyperparameter("input3", [0, 1])
-        hp4 = Constant("Or", "True")
-        cond1 = EqualsCondition(hp4, hp1, 1)
+    hp1 = CategoricalHyperparameter("input1", [0, 1])
+    hp2 = CategoricalHyperparameter("input2", [0, 1])
+    hp3 = CategoricalHyperparameter("input3", [0, 1])
+    hp4 = Constant("And", "True")
+    cond1 = EqualsCondition(hp4, hp1, 1)
 
-        self.assertRaises(ValueError, OrConjunction, cond1)
+    # Only one condition in an AndConjunction!
+    with pytest.raises(ValueError):
+        AndConjunction(cond1)
 
-        cond2 = EqualsCondition(hp4, hp2, 1)
-        cond3 = EqualsCondition(hp4, hp3, 1)
+    cond2 = EqualsCondition(hp4, hp2, 1)
+    cond3 = EqualsCondition(hp4, hp3, 1)
 
-        andconj1 = OrConjunction(cond1, cond2)
-        andconj1_ = OrConjunction(cond1, cond2)
-        assert andconj1 == andconj1_
+    andconj1 = AndConjunction(cond1, cond2)
+    andconj1_ = AndConjunction(cond1, cond2)
+    assert andconj1 == andconj1_
 
-        # Test setting vector idx
-        hyperparameter_idx = {hp1.name: 0, hp2.name: 1, hp3.name: 2, hp4.name: 3}
-        andconj1.set_vector_idx(hyperparameter_idx)
-        assert andconj1.get_parents_vector() == [0, 1]
-        assert andconj1.get_children_vector() == [3, 3]
+    # Test setting vector idx
+    hyperparameter_idx = {hp1.name: 0, hp2.name: 1, hp3.name: 2, hp4.name: 3}
+    andconj1.set_vector_idx(hyperparameter_idx)
+    assert andconj1.get_parents_vector() == [0, 1]
+    assert andconj1.get_children_vector() == [3, 3]
 
-        andconj2 = OrConjunction(cond2, cond3)
-        assert andconj1 != andconj2
+    andconj2 = AndConjunction(cond2, cond3)
+    assert andconj1 != andconj2
 
-        andconj3 = OrConjunction(cond1, cond2, cond3)
-        assert str(andconj3) == "(Or | input1 == 1 || Or | input2 == 1 || Or | input3 == 1)"
+    andconj3 = AndConjunction(cond1, cond2, cond3)
+    assert str(andconj3) == "(And | input1 == 1 && And | input2 == 1 && And | input3 == 1)"
 
-    def test_nested_conjunctions(self):
-        hp1 = CategoricalHyperparameter("input1", [0, 1])
-        hp2 = CategoricalHyperparameter("input2", [0, 1])
-        hp3 = CategoricalHyperparameter("input3", [0, 1])
-        hp4 = CategoricalHyperparameter("input4", [0, 1])
-        hp5 = CategoricalHyperparameter("input5", [0, 1])
-        hp6 = Constant("AND", "True")
+    # Test __eq__
+    assert andconj1 != andconj3
+    assert andconj1 != "String"
 
-        cond1 = EqualsCondition(hp6, hp1, 1)
-        cond2 = EqualsCondition(hp6, hp2, 1)
-        cond3 = EqualsCondition(hp6, hp3, 1)
-        cond4 = EqualsCondition(hp6, hp4, 1)
-        cond5 = EqualsCondition(hp6, hp5, 1)
 
-        conj1 = AndConjunction(cond1, cond2)
-        conj2 = OrConjunction(conj1, cond3)
-        conj3 = AndConjunction(conj2, cond4, cond5)
+def test_or_conjunction():
+    with pytest.raises(TypeError):
+        OrConjunction("String1", "String2")
 
-        # TODO: this does not look nice, And should depend on a large
-        # conjunction, there should not be many ANDs inside this string!
-        assert (
-            str(conj3)
-            == "(((AND | input1 == 1 && AND | input2 == 1) || AND | input3 == 1) && AND | input4 == 1 && AND | input5 == 1)"
-        )
+    hp1 = CategoricalHyperparameter("input1", [0, 1])
+    hp2 = CategoricalHyperparameter("input2", [0, 1])
+    hp3 = CategoricalHyperparameter("input3", [0, 1])
+    hp4 = Constant("Or", "True")
+    cond1 = EqualsCondition(hp4, hp1, 1)
 
-    def test_all_components_have_the_same_child(self):
-        hp1 = CategoricalHyperparameter("input1", [0, 1])
-        hp2 = CategoricalHyperparameter("input2", [0, 1])
-        hp3 = CategoricalHyperparameter("input3", [0, 1])
-        hp4 = CategoricalHyperparameter("input4", [0, 1])
-        hp5 = CategoricalHyperparameter("input5", [0, 1])
-        hp6 = Constant("AND", "True")
+    with pytest.raises(ValueError):
+        OrConjunction(cond1)
 
-        cond1 = EqualsCondition(hp1, hp2, 1)
-        cond2 = EqualsCondition(hp1, hp3, 1)
-        cond3 = EqualsCondition(hp1, hp4, 1)
-        cond4 = EqualsCondition(hp6, hp4, 1)
-        cond5 = EqualsCondition(hp6, hp5, 1)
+    cond2 = EqualsCondition(hp4, hp2, 1)
+    cond3 = EqualsCondition(hp4, hp3, 1)
 
-        AndConjunction(cond1, cond2, cond3)
-        AndConjunction(cond4, cond5)
-        self.assertRaisesRegex(
-            ValueError,
-            "All Conjunctions and Conditions must have " "the same child.",
-            AndConjunction,
-            cond1,
-            cond4,
-        )
+    andconj1 = OrConjunction(cond1, cond2)
+    andconj1_ = OrConjunction(cond1, cond2)
+    assert andconj1 == andconj1_
 
-    def test_condition_from_cryptominisat(self):
-        parent = CategoricalHyperparameter("blkrest", ["0", "1"], default_value="1")
-        child = UniformIntegerHyperparameter("blkrestlen", 2000, 10000, log=True)
-        condition = EqualsCondition(child, parent, "1")
-        assert not condition.evaluate({"blkrest": "0"})
-        assert condition.evaluate({"blkrest": "1"})
+    # Test setting vector idx
+    hyperparameter_idx = {hp1.name: 0, hp2.name: 1, hp3.name: 2, hp4.name: 3}
+    andconj1.set_vector_idx(hyperparameter_idx)
+    assert andconj1.get_parents_vector() == [0, 1]
+    assert andconj1.get_children_vector() == [3, 3]
 
-    def test_get_parents(self):
-        # Necessary because we couldn't call cs.get_parents for
-        # clasp-sat-params-nat.pcs
-        counter = UniformIntegerHyperparameter("bump", 10, 4096, log=True)
-        _1_S_countercond = CategoricalHyperparameter("cony", ["yes", "no"])
-        _1_0_restarts = CategoricalHyperparameter(
-            "restarts",
-            ["F", "L", "D", "x", "+", "no"],
-            default_value="x",
-        )
+    andconj2 = OrConjunction(cond2, cond3)
+    assert andconj1 != andconj2
 
-        condition = EqualsCondition(counter, _1_S_countercond, "yes")
-        # All conditions inherit get_parents from abstractcondition
-        assert [_1_S_countercond] == condition.get_parents()
-        condition2 = InCondition(counter, _1_0_restarts, ["F", "D", "L", "x", "+"])
-        # All conjunctions inherit get_parents from abstractconjunction
-        conjunction = AndConjunction(condition, condition2)
-        assert [_1_S_countercond, _1_0_restarts] == conjunction.get_parents()
+    andconj3 = OrConjunction(cond1, cond2, cond3)
+    assert str(andconj3) == "(Or | input1 == 1 || Or | input2 == 1 || Or | input3 == 1)"
+
+
+def test_nested_conjunctions():
+    hp1 = CategoricalHyperparameter("input1", [0, 1])
+    hp2 = CategoricalHyperparameter("input2", [0, 1])
+    hp3 = CategoricalHyperparameter("input3", [0, 1])
+    hp4 = CategoricalHyperparameter("input4", [0, 1])
+    hp5 = CategoricalHyperparameter("input5", [0, 1])
+    hp6 = Constant("AND", "True")
+
+    cond1 = EqualsCondition(hp6, hp1, 1)
+    cond2 = EqualsCondition(hp6, hp2, 1)
+    cond3 = EqualsCondition(hp6, hp3, 1)
+    cond4 = EqualsCondition(hp6, hp4, 1)
+    cond5 = EqualsCondition(hp6, hp5, 1)
+
+    conj1 = AndConjunction(cond1, cond2)
+    conj2 = OrConjunction(conj1, cond3)
+    conj3 = AndConjunction(conj2, cond4, cond5)
+
+    # TODO: this does not look nice, And should depend on a large
+    # conjunction, there should not be many ANDs inside this string!
+    assert (
+        str(conj3)
+        == "(((AND | input1 == 1 && AND | input2 == 1) || AND | input3 == 1) && AND | input4 == 1 && AND | input5 == 1)"
+    )
+
+
+def test_all_components_have_the_same_child():
+    hp1 = CategoricalHyperparameter("input1", [0, 1])
+    hp2 = CategoricalHyperparameter("input2", [0, 1])
+    hp3 = CategoricalHyperparameter("input3", [0, 1])
+    hp4 = CategoricalHyperparameter("input4", [0, 1])
+    hp5 = CategoricalHyperparameter("input5", [0, 1])
+    hp6 = Constant("AND", "True")
+
+    cond1 = EqualsCondition(hp1, hp2, 1)
+    cond2 = EqualsCondition(hp1, hp3, 1)
+    cond3 = EqualsCondition(hp1, hp4, 1)
+    cond4 = EqualsCondition(hp6, hp4, 1)
+    cond5 = EqualsCondition(hp6, hp5, 1)
+
+    AndConjunction(cond1, cond2, cond3)
+    AndConjunction(cond4, cond5)
+    with pytest.raises(ValueError):
+        AndConjunction(cond1, cond4)
+
+
+def test_condition_from_cryptominisat():
+    parent = CategoricalHyperparameter("blkrest", ["0", "1"], default_value="1")
+    child = UniformIntegerHyperparameter("blkrestlen", 2000, 10000, log=True)
+    condition = EqualsCondition(child, parent, "1")
+    assert not condition.evaluate({"blkrest": "0"})
+    assert condition.evaluate({"blkrest": "1"})
+
+
+def test_get_parents():
+    # Necessary because we couldn't call cs.get_parents for
+    # clasp-sat-params-nat.pcs
+    counter = UniformIntegerHyperparameter("bump", 10, 4096, log=True)
+    _1_S_countercond = CategoricalHyperparameter("cony", ["yes", "no"])
+    _1_0_restarts = CategoricalHyperparameter(
+        "restarts",
+        ["F", "L", "D", "x", "+", "no"],
+        default_value="x",
+    )
+
+    condition = EqualsCondition(counter, _1_S_countercond, "yes")
+    # All conditions inherit get_parents from abstractcondition
+    assert [_1_S_countercond] == condition.get_parents()
+    condition2 = InCondition(counter, _1_0_restarts, ["F", "D", "L", "x", "+"])
+    # All conjunctions inherit get_parents from abstractconjunction
+    conjunction = AndConjunction(condition, condition2)
+    assert [_1_S_countercond, _1_0_restarts] == conjunction.get_parents()
